requirements.txt
Django~=3.1
djangorestframework~=3.12
tictactoe-py~=1.0
